Refactor prompts page structure
